requirements.txt
# This requirements file lists all dependecies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in this project's lib/ directory.  The `lib` directory is added to `sys.path`
# by `appengine_config.py`.
bottle==0.12.7
PyCrypto==2.6
base58==0.2.2
